+++ mne/io/fiff/tests/test_raw_fiff.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # Author: Alexandre Gramfort <alexandre.gramfort@telecom-paristech.fr>
 #         Denis Engemann <denis.engemann@gmail.com>
 #
@@ -53,7 +54,7 @@ def test_acq_skip():
     assert_equal(len(raw.times), 17000)
     annotations = raw.annotations
     assert_equal(len(annotations), 3)  # there are 3 skips
-    assert_allclose(annotations.onset, [2, 7, 11])
+    assert_allclose(annotations.onset, [14, 19, 23])
     assert_allclose(annotations.duration, [2., 2., 3.])  # inclusive!
     data, times = raw.get_data(
         picks, reject_by_annotation='omit', return_times=True)
@@ -398,12 +399,26 @@ def test_split_files():
 
     assert_allclose(raw_1.buffer_size_sec, 10., atol=1e-2)  # samp rate
     split_fname = op.join(tempdir, 'split_raw.fif')
+    # intended filenames
+    split_fname_elekta_part2 = op.join(tempdir, 'split_raw-1.fif')
+    split_fname_bids_part1 = op.join(tempdir, 'split_raw_part-01_meg.fif')
+    split_fname_bids_part2 = op.join(tempdir, 'split_raw_part-02_meg.fif')
     raw_1.set_annotations(Annotations([2.], [5.5], 'test'))
     raw_1.save(split_fname, buffer_size_sec=1.0, split_size='10MB')
 
+    # check that the filenames match the intended pattern
+    assert op.exists(split_fname_elekta_part2)
+    # check that filenames are being formatted correctly for BIDS
+    raw_1.save(split_fname,
+               buffer_size_sec=1.0, split_size='10MB',
+               split_naming='bids',
+               overwrite=True)
+    assert op.exists(split_fname_bids_part1)
+    assert op.exists(split_fname_bids_part2)
+
     raw_2 = read_raw_fif(split_fname)
     assert_allclose(raw_2.buffer_size_sec, 1., atol=1e-2)  # samp rate
-    assert_array_equal(raw_1.annotations.onset, raw_2.annotations.onset)
+    assert_array_almost_equal(raw_1.annotations.onset, raw_2.annotations.onset)
     assert_array_equal(raw_1.annotations.duration, raw_2.annotations.duration)
     assert_array_equal(raw_1.annotations.description,
                        raw_2.annotations.description)
@@ -411,6 +426,12 @@ def test_split_files():
     data_2, times_2 = raw_2[:, :]
     assert_array_equal(data_1, data_2)
     assert_array_equal(times_1, times_2)
+
+    with pytest.warns(RuntimeWarning, match='does not conform to MNE'):
+        raw_bids = read_raw_fif(split_fname_bids_part1)
+    data_bids, times_bids = raw_bids[:, :]
+    assert_array_equal(data_1, data_bids)
+    assert_array_equal(times_1, times_bids)
 
     # test the case where we only end up with one buffer to write
     # (GH#3210). These tests rely on writing meas info and annotations
@@ -513,16 +534,16 @@ def test_io_raw():
     rng = np.random.RandomState(0)
     tempdir = _TempDir()
     # test unicode io
-    for chars in [b'\xc3\xa4\xc3\xb6\xc3\xa9', b'a']:
+    for chars in [u'äöé', 'a']:
         with read_raw_fif(fif_fname) as r:
             assert ('Raw' in repr(r))
             assert (op.basename(fif_fname) in repr(r))
-            desc1 = r.info['description'] = chars.decode('utf-8')
+            r.info['description'] = chars
             temp_file = op.join(tempdir, 'raw.fif')
             r.save(temp_file, overwrite=True)
             with read_raw_fif(temp_file) as r2:
                 desc2 = r2.info['description']
-            assert_equal(desc1, desc2)
+            assert desc2 == chars
 
     # Let's construct a simple test for IO first
     raw = read_raw_fif(fif_fname).crop(0, 3.5)
@@ -1222,11 +1243,12 @@ def test_save():
                         raw.info['meas_date'][0] +
                         raw.first_samp / raw.info['sfreq'])
     raw.set_annotations(annot)
+    annot = raw.annotations
     new_fname = op.join(op.abspath(op.curdir), 'break_raw.fif')
     raw.save(op.join(tempdir, new_fname), overwrite=True)
     new_raw = read_raw_fif(op.join(tempdir, new_fname), preload=False)
     pytest.raises(ValueError, new_raw.save, new_fname)
-    assert_array_equal(annot.onset, new_raw.annotations.onset)
+    assert_array_almost_equal(annot.onset, new_raw.annotations.onset)
     assert_array_equal(annot.duration, new_raw.annotations.duration)
     assert_array_equal(annot.description, new_raw.annotations.description)
     assert_equal(annot.orig_time, new_raw.annotations.orig_time)
@@ -1247,7 +1269,7 @@ def test_annotation_crop():
     onsets = raw.annotations.onset
     durations = raw.annotations.duration
     # 2*5s clips combined with annotations at 2.5s + 2s clip, annotation at 1s
-    assert_array_almost_equal([2.5, 7.5, 11.], onsets[:3], decimal=2)
+    assert_array_almost_equal(onsets[:3], [47.95, 52.95, 56.46], decimal=2)
     assert_array_almost_equal([2., 2.5, 1.], durations[:3], decimal=2)
 
     # test annotation clipping